+++ internal/utils/utils.go
@@ -68,6 +68,37 @@ func ExtractDatabaseErrorMessage(response *http.Response) error {
 	return errors.New(responseObj.Message)
 }
 
+func ExtractMCErrorMessage(response *http.Response) error {
+	bodyBytes, err := io.ReadAll(response.Body)
+	if err != nil {
+		return nil
+	}
+
+	_ = response.Body.Close()
+
+	var responseObj []struct {
+		Message   string `json:"message"`
+		FieldName string `json:"fieldName"`
+	}
+
+	err = json.Unmarshal(bodyBytes, &responseObj)
+	if err != nil {
+		var responseMessage string
+		err = json.Unmarshal(bodyBytes, &responseMessage)
+		if err != nil {
+			return nil
+		}
+
+		return errors.New(responseMessage)
+	}
+
+	if len(responseObj) == 0 {
+		return nil
+	}
+
+	return errors.New(responseObj[0].Message)
+}
+
 func ExtractShieldErrorMessage(response *http.Response) error {
 	bodyBytes, err := io.ReadAll(response.Body)
 	if err != nil {